github_agent/pyproject.toml
[project]
name = "github-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.2",
    "mcp>=1.15.0",
    "openai>=1.109.1",
    "openai-agents>=0.3.2",
    "python-dotenv>=1.1.1",
]
